Add Nix shell
This is just for my personal setup. It's not required and unless you
intentionally use Nix/Direnv, you can just ignore it and it will do
nothing. But I find it nice to commit regardless as the `shell.nix` is a
good reference of what system-wide packages you'll need to work with
this project (in this case, just a recent `nodejs` version).