Add transformers
This will hold all the conversion logic, so going from (in this case
just JSON but it might have been DB) source to Typescript models, as
well as Typescript models into View models. Nice side effect of going
about it this way is that it keeps the functions pure, which makes them
super easy to unit test.