Implement GET /memberships route
In implementing this, I added a `MembershipsService` as well as a
controller. The service layer will take care of joining data and
whatever else from the repositories layer. The controller layer will be
basically our routes entrypoint.

With this commit, I also splitted the `index.ts` into a separate
`app.ts` that just takes care of initializing the app, while `index.ts`
takes care of serving it. This is because it allows me to then test the
whole API flow.

A test for the GET /memberships using Express/Supertest has been added.
The way the code is structured, we could also test the service and
controller layers, but in this case I didn't see much benefit over
testing the whole flow through the route.